term: kan structure of sigma
